Add files via upload
This is the Jupyter Notebook of the Bangalore's House Price Prediction model. This is the very interesting project since it uses different techniques for preprocessing the dataset. I personally got many errors and took a lot of time to complete it, but its very exciting to create models like these.